+++ ft_printf/ft_putp.c
@@ -6,13 +6,13 @@
 /*   By: sangyepa <sangyepa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/27 22:12:50 by sangyepa          #+#    #+#             */
-/*   Updated: 2023/07/01 00:02:36 by sangyepa         ###   ########.fr       */
+/*   Updated: 2023/07/10 18:49:47 by sangyepa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include	"ft_printf.h"
 
-void	ft_putp(unsigned long long nbr, int *result)
+void	ft_putp(uintptr_t nbr, int *result)
 {
 	char	n;
 	int		check;
